requirements.test.txt
--index-url https://pypi.python.org/simple/

coveralls
whoosh
django_tools
